+++ web/src/components/molecules/index.ts
@@ -1,3 +1,3 @@
 export * from './dropdown/dropdown'
 export * from './text-field/text-field'
-export * from './login-form'
+export * from '../organisms/login-form/login-form'
