+++ web/src/components/organisms/index.ts
@@ -1,5 +1,5 @@
-export * from './login-card'
-export * from './auth-sidebar'
-export * from './app-sidebar'
-export * from './app-top-bar'
+export * from './login-form/login-form'
+export * from './auth-sidebar/auth-sidebar'
+export * from './app-sidebar/app-sidebar'
+export * from './app-top-bar/app-top-bar'
 export * from './register-institution-form/register-institution-form'
