+++ web/src/components/pages/index.ts
@@ -1,2 +1,2 @@
-export * from './login-page'
+export * from './login/login-page'
 export * from './register-institution-page/register-institution-page'
