+++ web/src/components/templates/index.ts
@@ -1,1 +1,1 @@
-export {}
+export * from './app-shell/app-shell'
