+++ web/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './navigation/navigation'
+export * from './page-headers/page-headers'
