+++ web/src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './login-form.types'
-export * from './login-card.types'
 export * from './navigation'
