--- web/src/types/login-card.types.ts
@@ -1,5 +1,0 @@
-export type LoginCardProps = {
-  className?: string
-  title?: string
-  subtitle?: string
-}
